Fixes for some compiler warnings, by Chris Bell, KD6ZWR.

+++ src/db.c
@@ -1,5 +1,5 @@
 /* -*- c-basic-indent: 4; indent-tabs-mode: nil -*-
- * $Id: db.c,v 1.24 2002/05/06 23:04:59 we7u Exp $
+ * $Id: db.c,v 1.25 2002/05/08 19:38:34 we7u Exp $
  *
  * XASTIR, Amateur Station Tracking and Information Reporting
  * Copyright (C) 1999,2000  Frank Giannandrea
@@ -5391,7 +5391,7 @@ void extract_area(DataRow *p_station, char *data) {
         }
 
         val = 0;
-        if (isdigit(data[1]) && isdigit(data[2])) {
+        if (isdigit((int)data[1]) && isdigit((int)data[2])) {
             val = (10 * (data[1] - '0')) + (data[2] - '0');
         }
         else {
@@ -5402,7 +5402,7 @@ void extract_area(DataRow *p_station, char *data) {
         temp_area.sqrt_lat_off = val;
 
         val = 0;
-        if (isdigit(data[5]) && isdigit(data[6])) {
+        if (isdigit((int)data[5]) && isdigit((int)data[6])) {
             val = (10 * (data[5] - '0')) + (data[6] - '0');
         }
         else {
@@ -8889,7 +8889,7 @@ int Create_object_item_tx_string(DataRow *p_station, char *line, int line_length
     altitude[0] = '\0'; // Start with empty string
     if (strlen(p_station->altitude) != 0) {   // Altitude was entered (we only handle feet currently)
         // Need to check for all digits, and 1 to 6 digits
-        if (isdigit(p_station->altitude[0])) {
+        if (isdigit((int)p_station->altitude[0])) {
             // Must convert from meters to feet before transmitting
             temp2 = (int)( (atof(p_station->altitude) / 0.3048) + 0.5);
             if ( (temp2 >= 0) && (temp2 <= 999999) ) {
