+++ src/main.h
@@ -1,5 +1,5 @@
 /*
- * $Id: main.h,v 1.8 2002/04/26 17:16:19 we7u Exp $
+ * $Id: main.h,v 1.9 2002/05/08 19:38:36 we7u Exp $
  *
  * XASTIR, Amateur Station Tracking and Information Reporting
  * Copyright (C) 1999,2000  Frank Giannandrea
@@ -258,12 +258,12 @@ extern gid_t egid;
 #define DISABLE_SETUID_PRIVILEGE do { \
 seteuid(getuid()); \
 setegid(getgid()); \
-if (debug_level & 4) { fprintf(stderr, "Changing euid to %d and egid to %d\n", getuid(), getgid()); } \
+if (debug_level & 4) { fprintf(stderr, "Changing euid to %d and egid to %d\n", (int)getuid(), (int)getgid()); } \
 } while(0)
 #define ENABLE_SETUID_PRIVILEGE do { \
 seteuid(euid); \
 setegid(egid); \
-if (debug_level & 4) { fprintf(stderr, "Changing euid to %d and egid to %d\n", euid, egid); } \
+if (debug_level & 4) { fprintf(stderr, "Changing euid to %d and egid to %d\n", (int)euid, (int)egid); } \
 } while(0)
 
 #endif /* XASTIR_MAIN_H */
